build: bump Java from 11 to 17
We want to migrate the lib to KMP. Exposed is currently used for database implementation and is not supported in KMP, it's JVM only. A very solid choice for a SQLite database implementation with KMP support is SqlDelight. SqlDelight requires JDK 17 to be able to run the SqlDelight plugin and compiler, hence this change.